Site prêt (première version)

+++ src/Controller/SeanceController.php
@@ -20,9 +20,9 @@ class SeanceController extends AbstractController
 {
 
     /**
-     * @Route("/", name="app_seance_index", methods={"GET"})
+     * @Route("/", name="toutes_les_seances", methods={"GET"})
      */
-    public function index(EntityManagerInterface $entityManager, Request $request,SeanceRepository $seanceRepository, PaginatorInterface $paginator): Response
+    public function seanceIndex(EntityManagerInterface $entityManager, Request $request,SeanceRepository $seanceRepository, PaginatorInterface $paginator): Response
     {
 
             //La variable $données récupère toutes les séances tandis que $séances permet de paginer toutes les séances
@@ -39,9 +39,9 @@ class SeanceController extends AbstractController
     }
 
     /**
-     * @Route("/new", name="app_seance_new", methods={"GET", "POST"})
+     * @Route("/creerSeance", name="creation_seance", methods={"GET", "POST"})
      */
-    public function new(Request $request, SeanceRepository $seanceRepository): Response
+    public function creationSeance(Request $request, SeanceRepository $seanceRepository): Response
     {
 
         //La méthode getUser permet de récupérer des contenus de l'utilisateur connecté
@@ -63,7 +63,7 @@ class SeanceController extends AbstractController
                 return $this->redirectToRoute('accueil', [], Response::HTTP_SEE_OTHER);
             }
             
-            return $this->renderForm('seance/_form.html.twig', [
+            return $this->renderForm('seance/form_seance.html.twig', [
                 'form' => $form,
             ]);     
          }
@@ -74,47 +74,62 @@ class SeanceController extends AbstractController
     }
 
     /**
-     * @Route("/{id}", name="app_seance_show", methods={"GET"})
+     * @Route("/{id}", name="affiche_seance", methods={"GET"})
      */
-    public function show(Seance $seance): Response
+    public function afficheSeance(Seance $seance): Response
     {
-        return $this->render('seance/show.html.twig', [
+        return $this->render('seance/affiche_seance.html.twig', [
             'seance' => $seance,
         ]);
     }
 
     /**
-     * @Route("/{id}/edit", name="app_seance_edit", methods={"GET", "POST"})
+     * @Route("/{id}/modifier-seance", name="modifie_seance", methods={"GET", "POST"})
      */
-    public function edit(Request $request, Seance $seance, SeanceRepository $seanceRepository): Response
+    public function modifieSeance(Request $request, Seance $seance, SeanceRepository $seanceRepository): Response
     {
-        $form = $this->createForm(SeanceType::class, $seance);
-        $form->handleRequest($request);
 
-        if ($form->isSubmitted() && $form->isValid()) {
-            $seanceRepository->add($seance, true);
+        $user = $this->getUser();
 
-            $this->addFlash('success',"Séance mise à jour !");
-
+        if($user){
+            $form = $this->createForm(SeanceType::class, $seance);
+            $form->handleRequest($request);
+    
+            if ($form->isSubmitted() && $form->isValid()) {
+                $seanceRepository->add($seance, true);
+    
+                $this->addFlash('success',"Séance mise à jour !");
+    
+                return $this->redirectToRoute('accueil', [], Response::HTTP_SEE_OTHER);
+            }
+    
+            return $this->renderForm('seance/form_seance.html.twig', [
+                'seance' => $seance,
+                'form' => $form,
+            ]);
+        }
+        else{
             return $this->redirectToRoute('accueil', [], Response::HTTP_SEE_OTHER);
-        }
-
-        return $this->renderForm('seance/_form.html.twig', [
-            'seance' => $seance,
-            'form' => $form,
-        ]);
+         }
     }
 
     /**
-     * @Route("/suppresionSeance/{id}", name="app_seance_delete", methods={"GET", "POST"})
+     * @Route("/suppresionSeance/{id}", name="supprime_seance", methods={"GET", "POST"})
      */
-    public function delete(Request $request, Seance $seance, SeanceRepository $seanceRepository): Response
+    public function supprimeSeance(Request $request, Seance $seance, SeanceRepository $seanceRepository): Response
     {
-        if ($this->isCsrfTokenValid('delete'.$seance->getId(), $request->request->get('_token'))) {
-            $seanceRepository->remove($seance, true);
-        }
 
-        return $this->redirectToRoute('app_seance_index', [], Response::HTTP_SEE_OTHER);
+        $user = $this->getUser();
+
+        if($user){
+            if ($this->isCsrfTokenValid('delete'.$seance->getId(), $request->request->get('_token'))) {
+                $seanceRepository->remove($seance, true);
+            }
+            return $this->redirectToRoute('toutes_les_seances', [], Response::HTTP_SEE_OTHER);
+        }
+        else{
+            return $this->redirectToRoute('toutes_les_seances', [], Response::HTTP_SEE_OTHER);
+        }
     }
 
     /**
@@ -124,9 +139,15 @@ class SeanceController extends AbstractController
     {
         //La méthode getUser permet de récupérer des contenus de l'utilisateur connecté
         $user = $this->getUser();
-        $seance->addJoueur($user);
-        $seanceRepository->add($seance, true);
-        return $this->redirectToRoute('seance_participation_perso', [], Response::HTTP_SEE_OTHER);
+
+        if($user){
+            $seance->addJoueur($user);
+            $seanceRepository->add($seance, true);
+            return $this->redirectToRoute('seance_participation_perso', [], Response::HTTP_SEE_OTHER);
+        }
+        else{
+            return $this->redirectToRoute('accueil', [], Response::HTTP_SEE_OTHER);
+         }
     }
 
     /**
@@ -136,8 +157,13 @@ class SeanceController extends AbstractController
     {
         //La méthode getUser permet de récupérer des contenus de l'utilisateur connecté
         $user = $this->getUser();
-        $seance->removeJoueur($user);
-        $seanceRepository->add($seance, true);
-        return $this->redirectToRoute('seance_participation_perso', [], Response::HTTP_SEE_OTHER);
+        if($user){
+            $seance->removeJoueur($user);
+            $seanceRepository->add($seance, true);
+            return $this->redirectToRoute('seance_participation_perso', [], Response::HTTP_SEE_OTHER);
+        }
+        else{
+            return $this->redirectToRoute('accueil', [], Response::HTTP_SEE_OTHER);
+        }
     }
 }
